seperates my team into seperate component
